+++ src/main/kotlin/com/relogging/server/service/CrawlingService.kt
@@ -1,5 +1,5 @@
 package com.relogging.server.service
 
 interface CrawlingService {
-    fun crawlAndSaveNewsArticles()
+    fun crawlAndSaveNewsArticles(): Int
 }
